Fix reported style issues
FossilOrigin-Name: cc66fe5d13646c464505c701517ff13b1a4e199945abfacfe2c4286ac25a5e45

+++ mtg_proxy_printer/ui/page_scene.py
@@ -646,7 +646,7 @@ class PageScene(QGraphicsScene):
         )
 
     @staticmethod
-    def _mm_to_rounded_px(value: int) -> int:
+    def _mm_to_rounded_px(value: float) -> int:
         return round((value*unit_registry.mm*RESOLUTION).to("pixel").magnitude)
 
     def remove_cut_markers(self):
